# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AiUsageSummaryItem" : {
        "properties" : {
          "estimatedCostUsd" : {
            "type" : "number"
          },
          "feature" : {
            "type" : "string"
          },
          "inputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "model" : {
            "type" : "string"
          },
          "outputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requests" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "BucketSummary" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "percent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "percentUsed" : {
            "format" : "double",
            "type" : "number"
          },
          "remaining" : {
            "type" : "number"
          },
          "spent" : {
            "type" : "number"
          },
          "target" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "BudgetRequest" : {
        "properties" : {
          "amountLimit" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "month" : {
            "minLength" : 1,
            "pattern" : "\\d{4}-(0[1-9]|1[0-2])",
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amountLimit", "categoryId", "month" ],
        "type" : "object"
      },
      "BudgetRequestV2" : {
        "properties" : {
          "amountLimit" : {
            "format" : "int64",
            "maximum" : 99999999999999999,
            "minimum" : 1,
            "type" : "integer"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "month" : {
            "minLength" : 1,
            "pattern" : "\\d{4}-(0[1-9]|1[0-2])",
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amountLimit", "categoryId", "month" ],
        "type" : "object"
      },
      "BudgetResponse" : {
        "properties" : {
          "amountLimit" : {
            "type" : "number"
          },
          "amountLimitInBaseCurrency" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "month" : {
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "BudgetResponseV2" : {
        "properties" : {
          "amountLimit" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amountLimitInBaseCurrency" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "month" : {
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "BudgetRuleSummaryResponse" : {
        "properties" : {
          "buckets" : {
            "items" : {
              "$ref" : "#/components/schemas/BucketSummary"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "monthlyIncome" : {
            "type" : "number"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "unassignedSpent" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "BudgetSummaryItem" : {
        "properties" : {
          "budgeted" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "percentUsed" : {
            "type" : "number"
          },
          "remaining" : {
            "type" : "number"
          },
          "spent" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryItemV2" : {
        "properties" : {
          "budgeted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "percentUsed" : {
            "type" : "number"
          },
          "remaining" : {
            "format" : "int64",
            "type" : "integer"
          },
          "spent" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryResponse" : {
        "properties" : {
          "dailyAllowance" : {
            "type" : "number"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/BudgetSummaryItem"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "safeToSpend" : {
            "type" : "number"
          },
          "totalBudgeted" : {
            "type" : "number"
          },
          "totalSpent" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryResponseV2" : {
        "properties" : {
          "dailyAllowance" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/BudgetSummaryItemV2"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "safeToSpend" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalBudgeted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalSpent" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "CategoryRequest" : {
        "properties" : {
          "icon" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "CategoryResponse" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ExpenseRequest" : {
        "properties" : {
          "amount" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "category" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "currency" : {
            "maxLength" : 3,
            "minLength" : 0,
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 1.0E-6,
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amount", "description" ],
        "type" : "object"
      },
...
      "ExpenseResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "amountInBaseCurrency" : {
            "type" : "number"
          },
          "category" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ObservabilitySummary" : {
        "properties" : {
          "activeUsers24h" : {
            "format" : "int64",
            "type" : "integer"
          },
          "activeUsers7d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signups30d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signups7d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signupsToday" : {
            "format" : "int64",
            "type" : "integer"
          },
          "topUsers30d" : {
            "items" : {
              "$ref" : "#/components/schemas/TopUserItem"
            },
            "type" : "array"
          },
          "totalUsers" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RecurringExpenseRequest" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "amount" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "dayOfMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dayOfWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "frequency" : {
            "enum" : [ "MONTHLY", "WEEKLY", "YEARLY" ],
            "type" : "string"
          },
          "monthOfYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "amount", "frequency", "name" ],
        "type" : "object"
      },
...
      "RecurringExpenseResponse" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "amount" : {
            "type" : "number"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "dayOfMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dayOfWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "frequency" : {
            "enum" : [ "MONTHLY", "WEEKLY", "YEARLY" ],
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "monthOfYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TopUserItem" : {
        "properties" : {
          "requests" : {
            "format" : "int64",
            "type" : "integer"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Backend-issued JWT from POST /auth/login, sent as `Authorization: Bearer <token>`.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/admin/ai-usage/summary" : {
      "get" : {
        "operationId" : "summary_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AiUsageSummaryItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-ai-usage-controller" ]
      }
    },
...
    "/admin/observability/summary" : {
      "get" : {
        "operationId" : "summary_2",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ObservabilitySummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-observability-controller" ]
      }
    },
...
    "/api/v2/budgets" : {
      "delete" : {
        "operationId" : "v2DeleteAllBudgets",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2ListBudgets",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BudgetResponseV2"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      },
      "post" : {
        "operationId" : "v2CreateBudget",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequestV2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponseV2"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      }
    },
    "/api/v2/budgets/summary" : {
      "get" : {
        "operationId" : "v2BudgetSummary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetSummaryResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      }
    },
    "/api/v2/budgets/{id}" : {
      "delete" : {
        "operationId" : "v2DeleteBudget",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      },
      "put" : {
        "operationId" : "v2UpdateBudget",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequestV2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      }
    },
...
    "/budget-rules/summary" : {
      "get" : {
        "operationId" : "summary_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      }
    },
    "/budgets" : {
      "delete" : {
        "operationId" : "deleteAll_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "get" : {
        "operationId" : "list_4",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BudgetResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
    "/budgets/summary" : {
      "get" : {
        "operationId" : "summary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
    "/budgets/{id}" : {
      "delete" : {
        "operationId" : "delete_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "put" : {
        "operationId" : "update_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
    "/categories" : {
      "delete" : {
        "operationId" : "deleteAll_2",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "get" : {
        "operationId" : "list_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "post" : {
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      }
    },
...
    "/expenses" : {
      "delete" : {
        "operationId" : "deleteAll_1",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "get" : {
        "operationId" : "list_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
...
    "/recurring" : {
      "delete" : {
        "operationId" : "deleteAll",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      },
      "get" : {
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RecurringExpenseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecurringExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecurringExpenseResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}